auth_app/serializers/user_serializer.py
from rest_framework import serializers
from inventory.models.user import User
from rest_framework.validators import ValidationError


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
